rename sd.h/c to fat32.h/c. remove compile warning. add help infors.
