+++ lab7/include/gpio.h
@@ -1,7 +1,4 @@
-//#define KVA 0xffff000000000000
-//#define MMIO_BASE (KVA + 0x3f000000)
-
-#define MMIO_BASE       0x3F000000
+# include "mem_addr.h"
 
 #define GPFSEL0         ((volatile unsigned int*)(MMIO_BASE+0x00200000))
 #define GPFSEL1         ((volatile unsigned int*)(MMIO_BASE+0x00200004))
