+++ lab7/rootfs/help/default
@@ -1,15 +1,27 @@
-Command    | Description
------------|--------------------------------------------
-hello      | print Hello World
-help       | print all avaiable commands
-reboot     | reboot pi
-buddy      | use "help buddy" for detail
-dma        | use "help dma" for detail
-mem        | use "mem status" to list memory alloc list
-alloc      | "alloc <bytes>"
-free       | "free <0x addr>"
-ls         | list all cpio files
-cat        | "cat <file name>"
-exec       | "exec <image file>"
-timer      | use "help timer" for detail
-setTimeout | setTimeout <MESSAGE> <SECONDS>
+Command       | Description
+--------------|-----------------------------------------------
+hello         | print Hello World
+help          | print all avaiable commands
+reboot        | reboot pi
+cpio          | use "help cpio" for detail
+buddy         | use "help buddy" for detail
+alloc         | use "help buddy" for detail
+free          | use "help buddy" for detail
+dma           | use "help dma" for detail
+mem           | use "mem status" to list memory alloc list
+alloc         | "alloc <bytes>"
+free          | "free <0x addr>"
+exec          | "exec <image file>"
+timer         | use "help timer" for detail
+setTimeout    | setTimeout <MESSAGE> <SECONDS>
+logger        | use "logger status" to show current log state
+logger <state>| state can be "fine", "info", "warning", "severe"
+tree          | show all vfs's files
+ls            | "ls <path>"
+cd            | "cd <path>"
+mkdir         | "mkdir <new dir name>"
+cat           | "cat <file path>"
+rm            | "cat <file path>"
+mount         | "mount <dir name> <fs name>"
+unmount       | "mount <mount point's dir name>"
+sdblock <num> | Show specific sd block
